addressing chris and alyssa pr comments; doc string typos, slight changes in tests, verbosity checking, pylint suppressing in experiment.py, deleting unncessary .dat file
