+++ smartsim/experiment.py
@@ -24,6 +24,8 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+# pylint: disable=too-many-lines
+
 import os
 import os.path as osp
 import typing as t
@@ -888,9 +890,9 @@ class Experiment:
         :param verbosity_level: verbosity level specified by user, defaults to info.
         :type verbosity_level: str
         :param output_format: Set output format. The possible accepted
-            output formats are `json`, `xml`, `html`, `plain_text`, `color_text`.
+            output formats are 'plain_text'.
             Defaults to 'plain_text'.
-        :type output_type: str
+        :type output_format: str
         :param output_filename: Specify name of file and extension to write
             preview data to. If no output filename is set, the preview will be
             output to stdout. Defaults to None.
